% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format_Summary.R
\name{Check_Sums}
\alias{Check_Sums}
\title{Format Summary}
\usage{
Check_Sums(
  input_file,
  Genome_Build,
  Pop,
  cols = c(BETA = "BETA", SE = "SE", Z = NULL, MAF = "MAF", Pval = "P", CHR = "CHR", POS
    = "POS", REF = "Allele1", ALT = "Allele2", SNP = "SNPID", n = "N", n_eff = NULL,
    n_case = NULL, n_ctrl = NULL),
  Spcf_n = NULL,
  Spcf_n_eff = NULL,
  Spcf_n_case = NULL,
  Spcf_n_ctrl = NULL,
  phenotype = "binary",
  allele_flip = TRUE,
  filter_by_hapmap3 = TRUE,
  fill_missing = TRUE,
  hapmap3_only = TRUE,
  minimum.P = 9.88131291682493e-324,
  save_path = NULL
)
}
\arguments{
\item{input_file}{Path of Summary statistics to be formatted}

\item{Genome_Build}{Genome build. hg37 or hg38}

\item{Pop}{Population group (EUR or EAS)}

\item{cols}{Column names of Summary statistics}

\item{Spcf_n}{Specify sample size if not stated in column}

\item{Spcf_n_eff}{Specify effective sample size if not stated in column}

\item{Spcf_n_case}{Size of overlapped samples if not stated in column}

\item{Spcf_n_ctrl}{Effective Sample size of summary to be adjusted}

\item{phenotype}{Effective Sample Size of the overlapped}

\item{allele_flip}{Sample Size of the all cases}

\item{filter_by_hapmap3}{Sample Size of the all controls}

\item{fill_missing}{Imputation missing columns with given Summary statistics}

\item{hapmap3_only}{Scope variants into hapmap3 markers only}

\item{minimum.P}{Minimum pvalue supported : default=9.88e-324}

\item{save_path}{Path for saving formatted summary statistics (optional)}
}
\value{
Formatted Summary Statistics
}
\description{
Reformat GWAS summary statistics file before overlap_adjustment
}
\examples{
cs = Check_Sums(system.file('extdata/example/consortium.ss',package = 'OAPRS'),
Genome_Build = "hg37", Pop = "eas",
cols = c(BETA="beta",Pval="pval",CHR="chrom",POS="pos",REF="ref",ALT="alt",SNP="rsids"),
Spcf_n=249625,Spcf_n_case = 50466, Spcf_n_ctrl = 199159)
}
