% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diagnostic_Plot.R
\name{diagnostic_plt}
\alias{diagnostic_plt}
\title{Score Evaluation}
\usage{
diagnostic_plt(
  scores,
  title,
  Output_Plot_path = NULL,
  keep.ind = NULL,
  ref.ind = NULL,
  ref_name = "",
  pheno_col = NULL,
  covar_cols = NULL,
  method_names = NULL
)
}
\arguments{
\item{scores}{Evaluated personalized scores}

\item{title}{Plot title}

\item{Output_Plot_path}{Output Path for generated diagnostic plot}

\item{keep.ind}{Sample indices to be evaluated.}

\item{ref.ind}{Sample indices for Reference group}

\item{ref_name}{Reference group name}

\item{pheno_col}{Column name for phenotype (Optional)}

\item{covar_cols}{Specify PRS method platform (prscs, ldpred2)}

\item{method_names}{New labels for methods (Optional)}
}
\value{
ggplot2 object of visual diagnostic
}
\description{
After applying polygenic risk score estimation tools, score evaluation for generating diagnostic plots
}
