% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OAPRS_0_2.r
\name{exclude_overlap}
\alias{exclude_overlap}
\title{Overlap Adjustment}
\usage{
exclude_overlap(
  file_all,
  file_ov,
  output_file,
  col_all = c("BETA", "SE", "AF_Allele2", "CHR", "POS", "Allele1", "Allele2", "P"),
  col_ov = c("BETA", "SE", "AF_Allele2", "CHR", "POS", "Allele1", "Allele2", "P"),
  n_all,
  n_ov,
  eff_all = NULL,
  eff_ov = NULL,
  n_case_all = NULL,
  n_ctrl_all = NULL,
  n_case_ov = NULL,
  n_ctrl_ov = NULL,
  phenotype,
  dropna = T,
  MAF = "snpinfo_1kg_hm3",
  rsid = NULL
)
}
\arguments{
\item{file_all}{Path of Summary statistics to be adjusted}

\item{file_ov}{Path of Summary statistics of Overlapped samples}

\item{output_file}{Output path of adjusted summary statistics}

\item{col_all}{Column names of Summary statistics to be adjusted}

\item{col_ov}{Column names of Summary statistics of Overlapped samples}

\item{n_all}{Sample size of summary to be adjusted}

\item{n_ov}{Size of overlapped samples}

\item{eff_all}{Effective Sample size of summary to be adjusted}

\item{eff_ov}{Effective Sample Size of the overlapped}

\item{n_case_all}{Sample Size of the all cases}

\item{n_ctrl_all}{Sample Size of the all controls}

\item{n_case_ov}{Sample Size of the overlapped cases}

\item{n_ctrl_ov}{Sample Size of the overlapped controls}

\item{phenotype}{Phenotype if binary : "binary", continuous : "continuous"}

\item{dropna}{Drop markers that has standard error NA's?}

\item{MAF}{external maf information path}

\item{rsid}{RS snp Marker column}
}
\value{
Adjusted summary
}
\description{
Overlap Adjustment from a GWAS summary with summary of Overlapped samples
}
\examples{
exclude_overlap("all_new.txt","/media/leelabsg-storage0/seokho/UKBB/overlapGWAS/Agen_T2D/KoGES_DM.tsv.gz","out_new_1004.txt",n_all = 433540,n_ov = 56918,
col_all = c("Beta","SE","EAF","Chr","Pos","Allele1","Allele2","P"),
eff_all =  211793, eff_ov = 10879.08,
col_ov = c("beta","sebeta","maf","chrom","pos","ref","alt","pval"),phenotype = "binary",dropna=T,
MAF = '/media/leelabsg-storage0/seokho/hapmap3/snpinfo_1kg_hm3');
temp2 <- F_to_C( c(50, 63, 23) );
}
