% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OAPRS_0_2.r
\name{Marker_select_ld}
\alias{Marker_select_ld}
\title{Overlap Adjustment}
\usage{
Marker_select_ld(
  input_file,
  outdir,
  pt = c(0.001, 1e-04, 1e-06, 1e-08),
  pvals = c(0, 0.01, 0.05, 0.1, 0.2, 0.5),
  hm_ref,
  EraSOR = NULL
)
}
\arguments{
\item{input_file}{Path of Summary statistics to be pruned}

\item{outdir}{Output directory of adjusted summary statistics}

\item{pt}{pvalue vector for pruning}

\item{pvals}{pvalue vector for Thresholding}

\item{hm_ref}{reference list}
}
\value{
Output path of adjusted summary
}
\description{
Overlap Adjustment from a GWAS summary with summary of Overlapped samples
}
\examples{
exclude_overlap("all_new.txt","/media/leelabsg-storage0/seokho/UKBB/overlapGWAS/Agen_T2D/KoGES_DM.tsv.gz","out_new_1004.txt",n_all = 433540,n_ov = 56918,
col_all = c("Beta","SE","EAF","Chr","Pos","Allele1","Allele2","P"),
eff_all =  211793, eff_ov = 10879.08,
col_ov = c("beta","sebeta","maf","chrom","pos","ref","alt","pval"),phenotype = "binary",dropna=T,
MAF = '/media/leelabsg-storage0/seokho/hapmap3/snpinfo_1kg_hm3');
temp2 <- F_to_C( c(50, 63, 23) );
}
