% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OAPRS_0_2.r
\name{PtoZ}
\alias{PtoZ}
\title{P-value to z-score}
\usage{
PtoZ(P, bet)
}
\arguments{
\item{P}{P-values to convert}

\item{bet}{Signed vector of coefficients}
}
\description{
P-value to z-score transformation
}
