% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Score_Evaluation.R
\name{score_eval}
\alias{score_eval}
\title{Score Evaluation}
\usage{
score_eval(
  prs_res_paths,
  lds,
  target_path,
  output_score_path = NULL,
  platform = "prscs",
  bim_path = NULL,
  fam_path = NULL,
  pheno_path = NULL,
  ID_col,
  pheno_col,
  cl = NULL
)
}
\arguments{
\item{prs_res_paths}{directory path of PRS result(e.g. prscs)}

\item{lds}{Variant Filter results}

\item{target_path}{Plink binary file header of target/validation genotypes}

\item{output_score_path}{Path for storing output scores.(Optional)}

\item{platform}{Specify PRS method platform (prscs, ldpred2)}

\item{bim_path}{Optional target bim file path}

\item{fam_path}{Optional target fam file path}

\item{pheno_path}{Covariate file of target/validation samples for additional}

\item{ID_col}{Column name for Sample ID in pheno_file}

\item{pheno_col}{Column name for phenotypes. If covariate file is not given. phenotype column from fam will be taken.}

\item{cl}{Cluster given from parallel}
}
\value{
Formatted Summary Statistics
}
\description{
After applying polygenic risk score estimation tools, score evaluation for generating diagnostic plots
}
