% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scorebed}
\alias{scorebed}
\title{Score Evaluation}
\usage{
scorebed(fileName, N, P, input, extract, keep)
}
\arguments{
\item{fileName}{directory path of PRS result(e.g. prscs)}

\item{N}{Number of samples (fam)}

\item{P}{Number of variants (bim)}

\item{input}{Weight matrix for multiplication}

\item{extract}{logical vector indicating variants to be scored}

\item{keep}{logical vector indicating samples to be scored}
}
\description{
After applying polygenic risk score estimation tools, score evaluation for generating diagnostic plots
}
\keyword{internal}
