% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overlap_Adjustment.r
\name{exclude_overlap}
\alias{exclude_overlap}
\title{Overlap Adjustment}
\usage{
exclude_overlap(file_all, file_ov, save_path = NULL, phenotype, dropna = T)
}
\arguments{
\item{file_all}{object/Path of Summary statistics to be adjusted}

\item{file_ov}{object/Path of Summary statistics of Overlapped samples}

\item{save_path}{Output path of adjusted summary statistics}

\item{phenotype}{Phenotype if binary : "binary", continuous : "continuous"}

\item{dropna}{Drop markers that has standard error NA's?}
}
\value{
Adjusted summary
}
\description{
Overlap Adjustment from a GWAS summary with summary of Overlapped samples
}
\examples{
print('adj_ss = exclude_overlap(cs,ts,"adj.txt",phenotype="binary")')
}
