% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marker_select_ld.R
\name{Marker_select_ld}
\alias{Marker_select_ld}
\title{Variant filtering for generating diagnostic plots}
\usage{
Marker_select_ld(
  ss,
  pt = 1e-04,
  probs = c(0, 0.01, 0.05, 0.1, 0.2, 0.5),
  Genome_Build,
  Pop,
  methods = c("all", "IVW", "RZ")
)
}
\arguments{
\item{ss}{Summary statistic object or path for formatted summary statistics}

\item{pt}{Threshold for pruning}

\item{probs}{Multiple P-value cutoffs for diagnostic plot}

\item{Genome_Build}{Genome build. hg37 or hg38}

\item{Pop}{Population group (EUR or EAS)}

\item{methods}{Specify sample size if not stated in column}
}
\value{
Dataframe of logical filter of markers, threshold, methods
}
\description{
Variant filtering by Independent LD blocks
}
